Use WebC for the page, instead of a layout
I discovered that there seems to be a bug using WebC layouts: the results of
render functions are not passed back through WebC, so the custom elements for
WebC components remain unchanged. Therefore, the gallery listing has to be
included in a page template, not a layout for the home page in order to get the
card component to be processed.